# Conflicts:
#	include/KeyboardTask.h
#	include/Packets/ACK.h
#	include/Packets/BCAST.h
#	include/Packets/DATA.h
#	include/Packets/DELRQ.h
#	include/Packets/DIRQ.h
#	include/Packets/DISC.h
#	include/Packets/ERROR.h
#	include/Packets/LOGRQ.h
#	include/connectionHandler.h
#	src/KeyboardTask.cpp
#	src/Packets/ACK.cpp
#	src/connectionHandler.cpp

+++ src/TFTPClient.cpp
@@ -17,7 +17,7 @@ int main (int argc, char *argv[]) {
         return -1;
     }
     std::string host = argv[1];
-    short port = atoi(argv[2]);
+    short port = (short)atoi(argv[2]);
 
     ConnectionHandler* connectionHandler= new ConnectionHandler(host, port);
     if (!connectionHandler->connect()) {
